==> notebooks/01_data_collection.ipynb <==
# **Data Collection**

## **Combining dataset**
Do yêu cầu dataset phải có tối thiểu 1000 dòng $\rightarrow$ nhóm quyết định gộp 2 file `.csv` trong tập dataset **Student Alcohol Consumption** gồm `student-mat.csv` và `student-por.csv` (2 file đều có chung các thuộc tính) thành `student-combine.csv` (có thêm thuộc tính `subject` cho `mat` và `por`) để xử lý cho các bước kế tiếp

import pandas as pd

df_mat = pd.read_csv("../data/raw/student-mat.csv")
df_por = pd.read_csv("../data/raw/student-por.csv")

# Thêm giá trị cho thuộc tính subject
df_mat['subject'] = 'mat'
df_por['subject'] = 'por'

# Gộp 2 data frame
df_combine = pd.concat([df_mat, df_por], axis=0, ignore_index=True)

df_combine.to_csv("../data/processed/student-clean.csv", index=False)

## **What subject data about?**
**Describe:**
- Tập dữ liệu mô tả kết quả học tập môn Toán và Bồ Đào Nha của học sinh trung học.
- Mục tiêu của bộ dữ liệu là hỗ trợ phân tích hành vi học sinh, thực hiện EDA và xây dựng các mô hình dự đoán điểm cuối kỳ (G3)

**Represent:**
- Bộ dữ liệu được thu thập từ một cuộc khảo sát học sinh trung học, tập trung vào hai môn học chính Toán và Ngữ văn Bồ Đào Nha, phản ánh đặc điểm thực tế của học sinh, các yếu tố trường học và hành vi 

## **What is the source of data**
**Platform name:** Kaggle

- [Student Alcohol Consumption](https://www.kaggle.com/datasets/uciml/student-alcohol-consumption?brid=ueCcCQEcoQEkZyTEcvMIgQ&fbclid=IwY2xjawOiFO1leHRuA2FlbQIxMQBzcnRjBmFwcF9pZAEwAAEes8Ocu3dLJp1fOuwlu3Um1QAqdmfHHI0ngGnSny2ij-cbD0l6SI88ejTbyOQ_aem_Koi5U_0BPkuJEqFzPD-FFA)

**Original authors / citation:**

P. Cortez and A. Silva. Using Data Mining to Predict Secondary School Student Performance. In A. Brito and J. Teixeira Eds., Proceedings of 5th FUture BUsiness TEChnology Conference (FUBUTEC 2008) pp. 5-12, Porto, Portugal, April, 2008, EUROSIS, ISBN 978-9077381-39-7.

Fabio Pagnotta, Hossain Mohammad Amran.

Email:fabio.pagnotta@studenti.unicam.it, mohammadamra.hossain '@' studenti.unicam.it
University Of Camerino

## **Is this data licensed for use**

Dữ liệu được cấp phép theo giấy phép [Open Data Commons Open Database License (ODbL)](https://opendatacommons.org/licenses/odbl/1.0/). Giấy phép này cho phép người dùng tự do sử dụng, chia sẻ và sửa đổi dữ liệu, với điều kiện rằng họ phải ghi nhận nguồn gốc của dữ liệu và chia sẻ các phiên bản sửa đổi dưới cùng một giấy phép.


## **How was this data collected**
Dữ liệu được thu thập từ một nghiên cứu về hành vi tiêu thụ rượu của học sinh trung học ở Bồ Đào Nha. Dữ liệu bao gồm thông tin về đặc điểm cá nhân, học tập và hành vi tiêu thụ rượu của học sinh. Dữ liệu được thu thập thông qua bảng câu hỏi tự báo cáo, nơi học sinh cung cấp thông tin về bản thân và thói quen uống rượu của họ. Nghiên cứu này nhằm mục đích hiểu rõ hơn về các yếu tố ảnh hưởng đến việc tiêu thụ rượu trong giới trẻ và các tác động tiềm ẩn của nó đối với thành tích học tập và sức khỏe.

## **Why choose this dataset?**
- Nhóm chọn dataset **Student Alcohol Consumption** vì tính thực tiễn và dữ liệu đa dạng. Dataset giúp người đọc có một cái nhìn khách quan về các yếu tố tác động, từ môi trường đến hoạt động hàng ngày, có ảnh hưởng như thế nào đến hành vi tiêu thụ cồn. Từ đó, nhóm có thể phân tích và đánh giá tác động thực tế của lối sống lên hiệu suất học tập (môn toán và môn tiếng Bồ Đào Nha) của học sinh trung học.

- Một số câu hỏi và insight tiềm năng có thể tìm thấy ở dataset này:
    1. Câu hỏi về hiệu suất học tập: mức độ tiêu thụ cồn (hàng ngày và cuối tuần) có mối liên hệ như thế nào đến kết quả học tập hay không?

    2. Câu hỏi về môi trường và xã hội: liệu học sinh dành nhiều thời gian đi chơi, hay ở trong một mối quan hệ tình cảm thường sẽ tiêu thụ cồn nhiều hơn hay không?

    3. Câu hỏi về gia đình: trình độ học vấn của ba, mẹ và sự ổn định trong gia đình có ảnh hưởng đến hành vi uống rượu của con cái hay không?

    4. Câu hỏi về sức khoẻ, đi học: tình trạng sức khoẻ hay số ngày nghỉ học có phản ánh sớm việc sa sút trong học tập?
